perf(frontend): 单据列表支持 ctrl 新开页面查看单据详情 #8852 # Reviewed, transaction id: 27927
