+++ dbm-ui/frontend/src/views/ticket-center/business/components/list/components/TableMode.vue
@@ -31,7 +31,7 @@
               v-if="data"
               text
               theme="primary"
-              @click="() => handleShowDetail(data)">
+              @click="(event: MouseEvent) => handleGoDetail(data, event)">
               {{ data.id }}
             </BkButton>
           </template>
@@ -57,16 +57,17 @@
   </div>
 </template>
 <script setup lang="ts">
-  import { getCurrentInstance, onActivated, onDeactivated } from 'vue';
+  import { onActivated } from 'vue';
   import { useI18n } from 'vue-i18n';
-  import { useRoute, useRouter } from 'vue-router';
+  import { useRoute } from 'vue-router';
 
   import TicketModel from '@services/model/ticket/ticket';
   import { getTickets } from '@services/source/ticket';
 
-  import { useStretchLayout, useUrlSearch } from '@hooks';
+  import { useUrlSearch } from '@hooks';
 
   import useDatePicker from '@views/ticket-center/common/hooks/use-date-picker';
+  import useOpenDetail from '@views/ticket-center/common/hooks/use-open-detail';
   import useSearchSelect from '@views/ticket-center/common/hooks/use-search-select';
   import TableModeTable from '@views/ticket-center/common/TableModeTable.vue';
   import TicketClone from '@views/ticket-center/common/TicketClone.vue';
@@ -74,19 +75,18 @@
 
   type IRowData = TicketModel;
 
-  const router = useRouter();
   const route = useRoute();
   const { t } = useI18n();
-  const currentInstance = getCurrentInstance();
 
-  const { getSearchParams, removeSearchParam } = useUrlSearch();
-  const { splitScreen: stretchLayoutSplitScreen } = useStretchLayout();
+  const { removeSearchParam } = useUrlSearch();
 
   const { value: datePickerValue, shortcutsRange } = useDatePicker();
 
   const { value: searachSelectValue, searchSelectData } = useSearchSelect({
     exclude: ['bk_biz_id'],
   });
+
+  const handleGoDetail = useOpenDetail();
 
   const dataSource = (params: ServiceParameters<typeof getTickets>) =>
     getTickets({
@@ -98,28 +98,9 @@
 
   const rowClass = (params: TicketModel) => (params.id === selectTicketId.value ? 'select-row' : '');
 
-  const handleShowDetail = (data: IRowData) => {
-    stretchLayoutSplitScreen();
-    selectTicketId.value = data.id;
-  };
-
   onActivated(() => {
     selectTicketId.value = Number(route.query.selectId);
     removeSearchParam('selectId');
-  });
-
-  onDeactivated(() => {
-    setTimeout(() => {
-      if (currentInstance!.isUnmounted) {
-        return;
-      }
-      router.replace({
-        params: {
-          ticketId: selectTicketId.value,
-        },
-        query: getSearchParams(),
-      });
-    });
   });
 </script>
 <style lang="less">
